services/horizon: Batch Upsert Accounts and Trust Lines (#2073)
This commit changes `DatabaseProcessor` to insert/update (upsert)
accounts and trust lines in batches.

In #2004 we added `LedgerEntryChangeCache` that aims to decrease number
of DB updates: all the changes connected to a single ledger entry are
squashed into just one DB query. Even though it gives a nice performance
boost when large number of ops change a small number of ledger entries,
it turns out this is not enough. When many ledger entries are changed in
a new ledger, DB connection round trip time takes significant percentage
of time in overall ledger processing time.

The SQL query was borrowed [1] from stellar-core.

Known limitations:

* This code shouldn't be the final code that is released. The aim of
this PR is to deploy it to stg environment to check it's performance.
* This is adding upsert queries for accounts and trust lines only. These
types are the most common changes in the recent ledgers. The final
solution should add batch upsert and batch delete for all ledger
entries.
* There's a potential to improve the code: autogenerating upsert query
for any ledger entry type, better tests. Will be done in a separate PR.
* It removes a code that check a number of rows affected by a query.
Unfortunately, the performance of the current solution is too bad to
keep it.

[1] https://github.com/stellar/stellar-core/blob/21469f90da1eacc6845017a520e179afb3772e65/src/ledger/LedgerTxnAccountSQL.cpp#L306-L336