services/keystore: use cors handler like our other apps (#2080)
Use CORS handler we use in other apps that implements CORS more completely.

The frontend folks are saying their seeing issues with the CORS OPTIONS handshake with the keystore.

I think our implementation of CORS is missing a step because the keystore is returning a 401 on an OPTIONS request. An OPTIONS request shouldn't be hitting an auth handler. Own implementation is passing through to the other middleware and the final handler of the endpoint and in those handler the 401 status code is being set when the user hasn't included credentials.

The implementation of CORS we use in our other apps won't pass through to the next handler for an OPTIONS request, which is how the preflight requests is supposed to be handled.