services/horizon: Fix tests mocking UpsertTrustLines (#2092)
Fixes a but in tests that are mocking `UpsertTrustLines` method.

Because all of the ledger entry changes are cached using
`LedgerEntryChangeCache` it's possible that the order of trustlines in
`UpsertTrustLines` will be different between calls if more than one
trust line is updated. `LedgerEntryChangeCache` is using map inside
which, when iterating, returns elements in random order.