exp/ingest/io: Add support for ingesting txmeta v2 (#2096)
This commit adds support for ingesting `TransactionMeta` when `V=2` and
updates XDR files.

As explained by @MonsieurNicolas in #1902.

> There is some missing meta in protocol version older than version 10.
This impacts transactions with one time signers (that get removed after
operations in v9 and earlier; and during signature verification in v10
and later). Those changes are invisible in the meta which may confuse
certain ingestion processors that keep track of signers on accounts.