Inline validateStartLedger() into ingestHistoryRange() (#2192)
In the init state it’s possible that during transition to ingestHistoryRangeState another horizon instance processed init and goes to the same state. Both horizon instances will not be able to ingest at the same time because the select for update lock is acquired at the beginning of ingestHistoryRange().

However, after the first horizon instance completes ingestHistoryRangeState it will release the lock and the other horizon instance will proceed. The validateStartLedger check will ensure that the horizon instance does not ingest the history range again. Instead it will exit early and go back to init. When that happens we should not report an error.

The previous code returned "expected latest history ledger to ..." as an error. But, it's actually not necessary to log this.