exp/services/webauth: make jwt issuer configurable (#2214)
### What
Add a parameter to configure the value of the JWT issuer that is
included in the `iss` claim of JWTs issued by the webauth service.

### Why
The issuer is hardcoded as the server signing key, but @JakeUrban
pointed out that the SEP-10 specification defines that the `iss` claim
be a URI. If it is a URI it needs to be configurable and set by the
people hosting the server and there is no need to prescribe the value
in the service so it is purely a configuration driven value now.

The change does not validate that the value is a URI because there is
no URI parsing code in the Go SDK, only URL parsing code. It's likely that
code would be sufficient in parsing the code but it would also
communicate that we expect the value to be a URL which is incorrect.
Given that there is little impact to parsing the value I've left it out.