keypair: add MustParseAddress, MustParseFull (#2216)
### What
Add MustParseAddress and MustParseFull functions that are the
panic-on-fail versions of ParseAddress and ParseFull.

### Why
It is useful to be able to use functions that do not return an error,
especially in tests when we are often parsing constant strings that we
know are valid and where a panic would very okay. It keeps tests more
succinct so that error conditions that won't happen are not noise.

I hope to use these functions in tests I am writing outside this package.