services/horizon/internal/expingest: Refactor ingestion system to use explicit FSM (#2220)
* Add missing invariant assertions in the beginning of resume and init

https://github.com/stellar/go/issues/2204 provides some details on scenarios where we can have multiple versions of ingestion running at the same time during a rolling deployment.

If we have checks against the ingestion version found in the db during the init and resume state we can detect these cases and shutdown.

* Refactor System to use explicit FSM

The ingestion system uses a finite state machine to model its behaviors.

In the current implementation the state machine is represented by single struct which represents all nodes in the state machine.

This PR replaces this representation with following interface:

type stateMachineNode interface {
	run(*System) (transition, error)
}

type transition struct {
	node          stateMachineNode
	sleepDuration time.Duration
}

Every node in the state machine is represented by a struct which implements the stateMachineNode interface.

There are two advantages to this approach:

All the state machine structs are immutable. The structs do not need to mutate their own fields (none of the structs have functions with pointer receivers) or fields in System .

The parameters for each node in the state machine are isolated. The previous implementation kept the parameters for all states in one state struct. In the old implementation it is possible for a state function to mistakenly use parameters that are not relevant to the sate.