keypair: remove unnecessary cast (#2255)
### What

Remove unnecessary cast of a `[]byte` to an `xdr.Signature` in the `keypair` packages `Sign` function.

### Why

The value is already a `[]byte` and doesn't need to be converted to an `xdr.Signature` and back to a `[]byte` again. The unnecessary casting makes it appear like the process of signing that is taking place here is specifically related to or controlled by the XDR package, but it isn't. The `Sign` function is a helper around taking a Stellar key and using it to perform standard ed25519 signing.