exp/services/recoverysigner: add account list endpoint (#2275)
### What

Add account list endpoint that returns a list of accounts the authenticate user has access to.

### Why

A user will need to be able to ask a recovery signer server what accounts are stored that it is allowed to recover. A user will use this endpoint to discover what accounts it has.