services/horizon: Fix docker-compose Postgres auth
By not pinning the Postgres version, authentication stopped working since now it
requires either setting the Postgres password or an explicit
`POSTGRES_HOST_AUTH_METHOD=trust`:

```
Error: Database is uninitialized and superuser password is not specified.
       You must specify POSTGRES_PASSWORD for the superuser. Use
       "-e POSTGRES_PASSWORD=password" to set it in "docker run".

       You may also use POSTGRES_HOST_AUTH_METHOD=trust to allow all connections
       without a password. This is *not* recommended. See PostgreSQL
       documentation about "trust":
       https://www.postgresql.org/docs/current/auth-trust.html
```

Making this breaking change on a patch release is arguably a mistake on
Postgres' end (see
https://github.com/docker-library/postgres/issues/681#issuecomment-586511283 )
but ... it is what it is. This commit sets `POSTGRES_HOST_AUTH_METHOD=trust` and
also pins the patch number of Postgres to make it future-proof.