exp/ingest: Polish interfaces (#2298)
This commit removes unused and improves existing interfaces in
`exp/ingest` subpackages:

* `StateReader` has been replaced by `ChangeReader`.
* `StateWriter` and `LedgerWriter` have been removed.
* Removed mocks for removed interfaces.
* Removed `PostgresTempSet`, made `memoryTempSet` struct and `tempSet`
interface private. Removed `tempSet` parameter in functions that create
internal state reader, `memoryTempSet` is used by default and can't be
changed.
* `Close` methods have been added to `ChangeReader` and `LedgerReader`.
* Moved some struct to it's own files (instead of `main.go`).

The motivation for adding `Close` methods is to make sure the interfaces
are future-proof. It's possible that the Stellar-Core ledger stream will
have to be closed (ex. tcp socket).