exp/services/webauth: fix the horizon timeout (#2312)
### What

Set the Horizon timeout to 1 minute for both the HTTP client and the horizonclient SDK.

### Why

While it looks like the same timeout is being used by the HTTP client and the horizonclient SDK it is not. The horizonclient SDK uses units of seconds and the HTTP client uses nanoseconds and the value being passed in to both is 60s. This causes the HTTP client to terminate most outgoing requests before they complete.