support/http: allow services to specify the logger used by the logging middleware (#2314)
### What

Allow services to specify the logger used by the logging middleware.

### Why

`LoggingMiddleware` gets the logger from the context using `"support/log".Ctx(...)`. If none is specified in the context the `"support/log".DefaultLogger` is used. Because the `LoggingMiddleware` is placed into the `chi.Mux` in the `NewMux` and `NewAPIMux` functions a service using these functions has no opportunity to set a logger on the context because a middleware cannot be inserted earlier. This means by using `NewMux` and `NewAPIMux` you are stuck with the default logger.

### Why Behind the Why

The reason I'm wanting this is because `webauth` and `recoverysigner` are similar to `ticker` in that they explicitly define all their dependencies and pass them down to the code using them as explicitly as possible. Because of this both of those applications define their own logger. Because we have some code like the LoggingMiddleware that magically pulls the logger out of the context setting it in the context is the only way to fully use a custom logger.