exp/services/recoverysigner: remove type, add identities (#2323)
### What
Remove the `type` parameter stored alongside registered accounts and replace it with exposing information about which identities are set on an account to clients.

Remove the `type` request and response parameters. Add the `identities` response parameter, and replace `owner_identities` and `other_identities` with `identities.owner` and `identities.other`.

### Why
The `type` existed purely for clients to get some sense of what purpose an account was registered, but it's really meta data about how a client is using the service and isn't necessarily a core concept for the
service.

The value was also synonymous with which identities are configured for an account. A `type` of `personal` will always have only an `owner` identity, and a `type` of `share` was always going to have an `other` identity alone or in addition to `owner`.

By removing the concept of `type` and instead just giving a client visibility into which identities are set on the account we reduce the number of concepts a client has to learn and know about which makes the API simpler, even if the change appears to make the API slightly more complex.

The reason I replaced `owner_identities` and `other_identities` is while thinking about the right way to expose the identities to the client it occurred to me the simplest method for a JavaScript client would be to have the identities be keys of an object, and so I structured the response that way, and it made sense to also modify the request to match.

### Known limitations

This is a breaking change to the API, but there are no consumers and the API is experimental.