exp/services/recoverysigner: add db migration using postgres (#2374)
### What

Add a `db migrations [up|down] [count]` command to recoverysigner that migrates the database up or down.

### Why

This is a first step to adding data persistence to recoverysigner and will allow us to make changes to the database schema and apply them incrementally. The structure of how the migrations work and are organized is based off what I see our other applications doing.

This is starting out with an empty schema because it's a well defined isolated chunk of functionality that is only focused on the mechanism for migrations and not the schema itself or the code that uses the database. I'll open subsequent PRs for those other things.

I've included functionality for migrating both up and down because it seems like we have that functionality in most of our applications, except Ticker. I prefer not to ship down migrations because I think in practice they are not useful and sometimes dangerous because they're often not tested well, they can destroy data, and production databases should always move forward in non-destructive ways. Down migrations can be useful in testing and also might be useful to others since this is open source and I assume the reason we've included them with most of our apps is due to a need. I'd like to put a guard in place that makes it a little harder to use the down migration so it isn't run accidentally in production but that isn't included yet. If you have any thoughts please let me know.

For #2339