exp/services/recoverysigner: make the db migrate invalid count error simpler (#2381)
### What
Change the error outputted when an invalid value is provided to the
recoverysigner's `db migrate [up|down] [count]` command as hte count
value from:
```
Invalid migration count, must be a number or not provided.
```
To:
```
Invalid migration count, must be a number.
```

### Why
The error message makes a suggestion to the user about something they
alternatively could do, remove the number, but this might not be what
they need to do and makes assumptions about what a user is doing. For
example, if the user saw this message and removed the invalid value and
were running a down migration it would be destructive. Separate to the
issue that it is too easy to migrate down all the way, this message goes
beyond just telling the user what is wrong and it's unnecessary for the
tool to go further than that.