exp/services/recoverysigner: update API to reflect SEP-30 v0.2.0 (#2383)
### What

Update API and related code to reflect the changes proposed in SEP-30 v0.2.0 in stellar/stellar-protocol#566.

### Why

To align the reference implementation with the SEP change being made to allow for easier expansion in the future of both additional roles and multiple values for each identity type (email, phone, etc).

I'm updating this code before finalizing the protocol change because I wanted to validate the protocol change in code. I'll merge both together.